+++ packages/presets/src/theme/tokens/index.ts
@@ -3,6 +3,8 @@ import { animations, easings } from './animations'
 import { colors } from './colors'
 import { lineHeights } from './line-heights'
 import { radii } from './radii'
+import { sizes } from './sizes'
+import { spacing } from './spacing'
 import { zIndex } from './z-index'
 
 export const tokens = defineTokens({
@@ -10,6 +12,8 @@ export const tokens = defineTokens({
   colors,
   easings,
   lineHeights,
+  sizes,
+  spacing,
   radii,
   zIndex,
 })