+++ packages/presets/src/theme/tokens/sizes.ts
@@ -1,0 +1,5 @@
+import { defineTokens } from '@pandacss/dev'
+
+export const sizes = defineTokens.sizes({
+  4.5: { value: '1.125rem' },
+})