+++ packages/presets/src/theme/tokens/spacing.ts
@@ -1,0 +1,5 @@
+import { defineTokens } from '@pandacss/dev'
+
+export const spacing = defineTokens.spacing({
+  4.5: { value: '1.125rem' },
+})